tools: Import the latest docker scripts

+++ tools/install-docker-lab.sh
@@ -1,13 +1,19 @@
 #!/bin/bash
 #
-# need to run with sudo
+# install-docker-lab.sh -- Build the docker image for the lab
+#
+# ref: https://docs.docker.com/engine/installation/linux/ubuntulinux/
 #
 
 TOP_DIR=$(dirname `readlink -f $0`)
 
-. ${TOP_DIR}/jekyll-config
+IMAGE=$(< $TOP_DIR/lab-name)
 
-image_name=$user/$url
+docker_without_sudo=0
+groups $USER | grep -q docker
+[ $? -eq 0 ] && docker_without_sudo=1
+
+# Make sure docker is installed
 which docker 2>&1 > /dev/null
 if [ $? -eq 1 ]; then
     sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
@@ -15,7 +21,7 @@ if [ $? -eq 1 ]; then
     sudo bash -c 'echo "deb https://apt.dockerproject.org/repo ubuntu-'${version}' main" > /etc/apt/sources.list.d/docker.list'
     sudo apt-get -y update
     sudo apt-get -y install apt-transport-https ca-certificates
-    sudo apt-get -y install docker-engine
+    sudo apt-get -y --force-yes install docker-engine
     sudo usermod -aG docker $USER
 
     # The --bip works like https://gist.github.com/ismell/6689836
@@ -25,18 +31,21 @@ if [ $? -eq 1 ]; then
 DOCKER_OPTS="\$DOCKER_OPTS --insecure-registry=registry.mirrors.aliyuncs.com"
 DOCKER_OPTS="\$DOCKER_OPTS --dns 8.8.8.8 --dns 8.8.4.4"
 DOCKER_OPTS="\$DOCKER_OPTS --bip=10.66.33.10/24"
+DOCKER_OPTS="\$DOCKER_OPTS --storage-opt dm.basesize=2G"
 # DOCKER HACK END 789527394722
 EOF'
+    # Restart to make sure the above opts work
+    sudo brctl delbr docker0
+    sudo service docker restart
 
 fi
 
-sudo docker build -t $image_name $TOP_DIR/
+# Build the lab
+sudo docker build -t $IMAGE $TOP_DIR/
 
-echo -e "\nNote: To let docker work without sudo, add $USER to docker group and restart the X session please.\n"
+# Note: Let docker without sudo, please restart X.
+[ $docker_without_sudo -eq 1 ] && exit 0
 
 sure='n'
-read -p '* Restart X to make sure docker work without sudo (Y/n): ' sure
-
-if [ "$sure" = 'Y' -o "$sure" = 'y' ]; then
-    sudo pkill X
-fi
+read -p 'LOG: Restart X to let docker work without sudo (Y/n): ' sure
+[ "$sure" = 'Y' -o "$sure" = 'y' ] && sudo pkill X
